pyproject.toml
[build-system]
requires = ["setuptools >= 75.0", "setuptools_scm[toml]", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nilmtk"
dynamic = ["version"]
dependencies = [
    "h5py",
    "hdf5plugin ",
    "hmmlearn~=0.3.0",
    "jupyterlab",
    "matplotlib",
    "networkx",
    #"nilm_metadata",
    "numpy",
    "pandas~=2.2.0",
    "scikit-learn",
    "scipy",
    "tables",
    "pyyaml",
]
requires-python = ">= 3.11"
authors = [{ name = "NILMTK developers" }]
description = "Estimate the energy consumed by individual appliances from whole-house power meter readings"
readme = "README.md"
license = { file = "LICENSE" }
keywords = [
    "smartmeters",
    "power",
    "electricity",
    "energy",
    "analytics",
    "redd",
    "disaggregation",
    "nilm",
    "nialm",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache 2.0",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Mathematics",
]

[project.optional-dependencies]
dev = ["coverage", "coveralls", "pytest", "ruff", "sphinx"]
full = ["deddiag-loader", "psycopg2"]

[project.urls]
Repository = "https://github.com/nilmtk/nilmtk"
Download = "https://github.com/nilmtk/nilmtk/tarball/master#egg=nilmtk-dev"

[tool.setuptools]
packages = ["nilmtk"]

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"
fallback_version = "0.0.1.dev0+unknown"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "nilmtk/datastore/tests",
    "nilmtk/preprocessing/tests",
    "nilmtk/stats/tests",
    "nilmtk/tests",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "nilmtk/legacy",
    "node_modules",
    "site-packages",
    "tests",
    "venv",
    "*.ipynb",
]
line-length = 120
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E731", "F403", "F405"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
exclude = ["docs/source/conf.py"]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false

[project.scripts]
generate_data = "nilmtk.tests.generate_data:create_all"
